feat(ZH-128): 로직 분리를 위한 custom hook 생성
